# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateAssignmentInput:
    properties:
      description:
        type: string
      due_date:
        type: string
      is_published:
        type: boolean
      max_score:
        type: integer
      title:
        type: string
    required:
    - due_date
    - max_score
    - title
    type: object
...
  dto.UpdateAssignmentInput:
    properties:
      description:
        type: string
      due_date:
        type: string
      is_published:
        type: boolean
      max_score:
        type: integer
      title:
        type: string
    required:
    - due_date
    - is_published
    - max_score
    - title
    type: object
...
info:
  contact: {}
  description: Backend API for Mini-LMS / Teacher Platform
  title: Teacher Platform API
  version: "1.0"
paths:
  /assignments/{id}:
    get:
      consumes:
      - application/json
      description: Get one assignment by id
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get assignment by id
      tags:
      - assignments
    patch:
      consumes:
      - application/json
      description: Update assignment by id
      parameters:
      - description: Assignment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update assignment input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateAssignmentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update assignment
      tags:
      - assignments
...
  /lessons/{id}/assignments:
    get:
      consumes:
      - application/json
      description: Get all assignments for a lesson
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get assignments by lesson
      tags:
      - assignments
    post:
      consumes:
      - application/json
      description: Create a new assignment in a lesson
      parameters:
      - description: Lesson ID
        in: path
        name: id
        required: true
        type: integer
      - description: Create assignment input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateAssignmentInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create assignment
      tags:
      - assignments
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"